Refactor: move chatroom dtos to separate package and improve semantics
