--- src/main/java/com/chat/chat_api/chatroom/CreateChat.java
@@ -1,5 +1,0 @@
-package com.chat.chat_api.chatroom;
-
-public record CreateChat(String chatName) {
-    
-}
