+++ src/main/java/com/chat/chat_api/chatroom/dto/CreateChatRequestDTO.java
@@ -1,0 +1,5 @@
+package com.chat.chat_api.chatroom.dto;
+
+public record CreateChatRequestDTO(String chatName) {
+    
+}
